+++ config/routes.rb
@@ -1,3 +1,14 @@
 Rails.application.routes.draw do
-  # For details on the DSL available within this file, see https://guides.rubyonrails.org/routing.html
+  root "users#profile"
+
+  get '/login', to: 'sessions#new'
+  post '/login', to: 'sessions#create'
+  post '/logout', to: 'sessions#destroy'
+  get '/signup', to: 'users#new'
+  get '/events/all', to: 'events#all'
+  post '/events/:id/attend', to: 'events#attend', as: 'attend'
+  post '/events/:id/unattend', to: 'events#unattend', as: 'unattend'
+
+  resources :users, except: [:index, :new]
+  resources :events, except: [:destroy, :edit, :update]
 end