+++ CHANGELOG.md
@@ -7,3 +7,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Authentication.
+- Create events.
+- View user events.
+- View all events.
+- Attend/Unattend events.
+- Separate attended from unattended events.
+- Separate upcoming from past events.
+- Used helpers for logic.
+- Testing models validations and associations.
+- Testing events management (creation, attending, and unattending).
+- Testing authentication (Login, and Logout).
+- Updated Readme file.
+- Moved logic test for signed in user out of view files
+- Use `shoulda-matchers` gem.